Convert devDepencies to an npm script
To ensure installation of latest packages, I opted to install them all
at once instead of leaving them in devDepencies and relying on
"npm update", since that only updates to minor versions, not the true
latest version.

I would've preferred to list them in an array and insert them into the
script with .join(" ") in a template literal, but json is presenting
difficulties.